requirements.txt
python-docx
openpyxl
pandas
requests
python-dotenv
openai
colorama
sentence-transformers
faiss-cpu
numpy
scikit-learn
docx2pdf
pdf2image
Pillow
PyMuPDF